norme ok, inversed iterative logic in check_infiles to save lines
